force

+++ custom_utils/GrammarWeight.java
@@ -10,10 +10,10 @@ import custom_utils.FileRead.*;
 
 public class GrammarWeight {
   void Parameters() {
-    idfConceptFact = new Double(100);
+    idfConceptFact = new Double(5);
     idfFreeWords = new Double(1);
 
-    tfConceptFact = new Double(10);
+    tfConceptFact = new Double(5);
     tfFreeWords = new Double(1);
 
   }
